LayoutTests/imported/w3c/web-platform-tests/referrer-policy/generic/inheritance/iframe-inheritance-srcdoc-child-expected.txt

Referrer Policy: iframes srcdoc child correctly inherit the ancestor's referrer policy


FAIL iframes srcdoc child correctly inherit the ancestor's referrer policy assert_equals: expected (string) "http://localhost:8800/" but got (undefined) undefined
